Migrate db utility functions from modifications to mpc_nodes module
